Issues #162 feat: AddGifticonUIModel 값 추가 및 Mapper 정리
