+++ presentation/src/main/java/com/lighthouse/presentation/mapper/GalleryImageMapper.kt
@@ -1,0 +1,12 @@
+package com.lighthouse.presentation.mapper
+
+import android.net.Uri
+import com.lighthouse.domain.model.GalleryImage
+import com.lighthouse.presentation.model.GalleryUIModel
+
+fun GalleryImage.toPresentation(index: Int = -1): GalleryUIModel.Gallery = GalleryUIModel.Gallery(
+    id = id,
+    uri = Uri.parse(contentUri),
+    selectedOrder = index,
+    date = date
+)
