+++ presentation/src/main/java/com/lighthouse/presentation/mapper/GifticonForAdditionMapper.kt
@@ -14,6 +14,7 @@ fun GifticonForAddition.toPresentation(id: Long): AddGifticonUIModel {
         hasImage = hasImage,
         name = name,
         brandName = brandName,
+        brandConfirm = false,
         barcode = barcode,
         expiredAt = expiredAt,
         isCashCard = isCashCard,
