+++ presentation/src/main/java/com/lighthouse/presentation/model/AddGifticonUIModel.kt
@@ -9,6 +9,7 @@ data class AddGifticonUIModel(
     val hasImage: Boolean,
     val name: String,
     val brandName: String,
+    val brandConfirm: Boolean,
     val barcode: String,
     val expiredAt: Date,
     val isCashCard: Boolean,
